requirements.txt
python-dotenv==1.0.0
psycopg2-binary==2.9.5
requests==2.28.0
pytz==2021.1
